sched/fair: Estimate capacity index ahead	energy computation
Based on the sched_group and waken task utilization, we can estimate
the capacity index ahead energy computation, and the capacity index is
consistent for all CPUs in the group.

This patch moves capacity index estimation ahead energy computation, and
optimize to only estimate capacity index once and remove redundant
estimations for every CPU in the code.

Change-Id: I634f2fa47d2e762a1b5e570c7f792b67e202a1d2
Signed-off-by: Leo Yan <leo.yan at linaro.org>